Crate problem118

Problem 118 (Pandigital prime sets)

Solution detail

Because of the unstructured way in which primes are distributed, any solution will need some element of brute-force. The key is in minimising the amount of repeated work.

One approach is to first, for each subset of the digits 1..9, check all permutations of those digits and work out how many of them are prime. Store that count in a HashMap keyed by the set of digits.

Then iterate over all ways of partitioning 1..9 into distinct subsets, and for each partition, calculate how many ways the digits can be arranged so that all items in the partition are prime. This is just the product of the number of permutations of each part that are prime, calculated earlier and stored in the HashMap.

Functions

answer
Solve the problem, returning the answer as a String

main

prime_permutations
Find the number of permutations of the given digits which are prime, using the given sieve for primality testing.

solve
Find the number of pandigital prime sets.

to_integer
Turn the given digits into a number